♻️ Move util/etc to `/shared`
