+++ shared/middleware/headers.js
